+++ lib/vor/lexer.ex
@@ -37,7 +37,8 @@ defmodule Vor.Lexer do
                    relation fact synthesize property never always eventually implies
                    forall where and or not in start_timer cancel_timer restart_timer
                    retransmit_last_response extern if else
-                   sends send broadcast system connect solve noop)
+                   sends send broadcast system connect solve noop
+                   map_get map_put map_has map_delete map_size map_sum map_merge)
 
     type = if name in keywords, do: :keyword, else: :identifier
     {rest, [{type, {line, col}, String.to_atom(name)}], context}